Commit 14
completed the send function, it work fine.
changed the code in the EcoServer to fit the column additionalInfo so that it will show years ,months and days if the book returned late, if the book returned before deadline , it will write in the column "Returned on time".
changed the methods that i work with to fit the changes